use explicit integral types instead of "int" in ComputeProgram::dispatch

+++ gl_classes/include/gl_classes/compute_program.h
@@ -3,6 +3,7 @@
 #include <utility>
 #include <iostream>
 #include <sstream>
+#include <cstdint>
 
 #include "gl_classes/program.h"
 
@@ -13,16 +14,17 @@ namespace gl_classes {
     public:
         ComputeProgram() : Program(){}
         virtual ~ComputeProgram(){}
-        virtual ComputeProgram& dispatch(int x, int y, int z, int gx, int gy, int gz)
+
+        virtual ComputeProgram& dispatch(uint64_t x, uint64_t y, uint64_t z, uint32_t gx, uint32_t gy, uint32_t gz)
         {
             return ComputeProgram::dispatch(
-                x/gx + ((x%gx == 0) ? 0 : 1),
-                y/gy + ((y%gy == 0) ? 0 : 1),
-                z/gz + ((z%gz == 0) ? 0 : 1)
+                static_cast<uint32_t>(x/gx + ((x%gx == 0) ? 0 : 1)),
+                static_cast<uint32_t>(y/gy + ((y%gy == 0) ? 0 : 1)),
+                static_cast<uint32_t>(z/gz + ((z%gz == 0) ? 0 : 1))
             );
         }
         
-        virtual ComputeProgram& dispatch(int x, int y, int z)
+        virtual ComputeProgram& dispatch(uint32_t x, uint32_t y, uint32_t z)
         {
             glDispatchCompute(x, y, z);
             checkGLError();
